Reduce max packet size. The plotter was blowing up when too many plot / curve connections were active because it ran out of heap.
Original Commit Rev: 2058 | 2016-10-17 18:38:00 -0500 (Mon, 17 Oct 2016)

+++ persistentParameters.cpp
@@ -56,13 +56,13 @@ void persistentParam_setPath(std::string path)
    }
    else
    {
-      std::string dir = fso::GetDir(path);
+      std::string dir = path;
       if(fso::DirExists(dir))
       {
          iniPath = dir + fso::dirSep() + PERSISTENT_APPEND;
       }
    }
-   if(fso::FileExists(iniPath) == false)
+   if(iniPath != "" && fso::FileExists(iniPath) == false)
    {
       std::string iniFileInit = "";
       fso::WriteFile(iniPath, iniFileInit);
